+++ src/main/java/com/jnulocker/auth/application/port/in/request/validation/ValidCodeValidator.java
@@ -1,4 +1,4 @@
-package com.jnulocker.auth.application.port.in.request.valid;
+package com.jnulocker.auth.application.port.in.request.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
